Add Event Listeners for each Button
I added an event listener for each button. I may try to optimize
it further once the calculator is complete. I also created a function
to create a string of numbers for each number press,to the left
or right variable depending on whether or not an operator was clicked.